Handle lists of colon separated records
Some commands provide lists of colon separated records, with each record
separated by a blank line. The records may be of the same or different
types. Allow commands to choose record types and fields based on the
contents of the record, and return lists of the records.

Also added the PAREN_PAIR_VAL value type for stuff like:

System Type : BMC (0xb)

With it, you can assign 'BMC' to one attribute and 0xb to another.